+++ telebot/src/main/java/exh3y/telebot/data/keyboards/ReplyKeyboardMarkup.java
@@ -1,4 +1,4 @@
-package exh3y.telebot.data;
+package exh3y.telebot.data.keyboards;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
